Add files via upload
Owing to the transcendent nature of time, COVID cases, even with the prevalence of substantial cure in forms of the vaccinations, appeared to have endured in some cases around the world since the release of the vaccines. 
This analysis is an earnest venture at capturing an updated exploratory analysis of COVID-19 cases in September 2023 on Microsoft SQL Server.